Move directives to directives/
