+++ package.json
@@ -14,6 +14,7 @@
     "/index.js.map",
     "/index.d.ts",
     "/lib/",
+    "/directives/",
     "/src/",
     "!/src/test/"
   ],
