+++ src/test/directives/repeat_test.ts
@@ -15,9 +15,9 @@
 /// <reference path="../../../node_modules/@types/mocha/index.d.ts" />
 /// <reference path="../../../node_modules/@types/chai/index.d.ts" />
 
+import {repeat} from '../../directives/repeat.js';
 import {html} from '../../index.js';
 import {render} from '../../lib/render.js';
-import {repeat} from '../../lib/repeat.js';
 import {stripExpressionDelimeters} from '../test-helpers.js';
 
 const assert = chai.assert;
